deployment/local-inference/tflite/bounding_box/requirements.txt
tensorflow==2.5.0
numpy==1.18.5
Pillow==7.2.0
matplotlib==3.1.3
opencv-python
tf-slim